+++ javax-kt/src/test/java/me/panpf/javaxkt/test/io/FileTest.kt
@@ -17,6 +17,9 @@
 package me.panpf.javaxkt.test.io
 
 import me.panpf.javaxkt.io.*
+import me.panpf.javaxkt.util.requireExist
+import me.panpf.javaxkt.util.requireIsDir
+import me.panpf.javaxkt.util.requireIsFile
 import org.junit.Assert
 import org.junit.Test
 import java.io.File