+++ javax/src/main/java/me/panpf/javax/util/Zipx.java
@@ -103,7 +103,7 @@ public class Zipx {
      * @throws IOException IO exceptions
      */
     public static File compressionDirTo(@NotNull final File sourceDirectory, @NotNull File destinationFile) throws IOException {
-        Filex.requireExist(sourceDirectory);
+        Premisex.requireExist(sourceDirectory);
         Premisex.require(sourceDirectory.isDirectory(), sourceDirectory.getPath() + " not directory");
         return compressionFilesTo(sourceDirectory.listFiles(), destinationFile, new Transformer<File, String>() {
             @NotNull
@@ -136,7 +136,7 @@ public class Zipx {
      */
     @NotNull
     public static File decompressionTo(@NotNull File zipSourceFile, @NotNull File destinationDir) throws IOException {
-        Filex.requireExist(zipSourceFile);
+        Premisex.requireExist(zipSourceFile);
         Premisex.require(!destinationDir.exists() || destinationDir.isDirectory(), destinationDir.getPath() + " not directory");
 
         ZipFile zipFile = new ZipFile(zipSourceFile);